Rename NodeVisitors NS to NodeVisitor
