+++ src/NodeVisitor/ReferencesAdder.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types = 1);
 
-namespace LanguageServer\NodeVisitors;
+namespace LanguageServer\NodeVisitor;
 
 use PhpParser\{NodeVisitorAbstract, Node};
 use LanguageServer\Protocol\{SymbolInformation, SymbolKind, Range, Position, Location};
